feat: updated feedbacks

+++ src/main/java/org/example/services/impl/MasterServiceImpl.java
@@ -2,7 +2,6 @@ package org.example.services.impl;
 
 
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.apache.tomcat.util.http.fileupload.FileUtils;
 import org.example.entities.*;
 
@@ -31,7 +30,6 @@ import java.nio.file.Paths;
 import java.util.*;
 
 
-@Slf4j
 @Service
 @RequiredArgsConstructor
 public class MasterServiceImpl implements MasterService {
@@ -148,21 +146,19 @@ public class MasterServiceImpl implements MasterService {
     @Override
     public String uploadPhoto(MultipartFile multipartFile, String username) throws IOException {
         Files.createDirectories(Paths.get(PATH_TO_MEDIA + username + "/photo/"));
-        String pathStr = PATH_TO_MEDIA + username + "/photo/" + multipartFile.getOriginalFilename();
-        File file = new File(pathStr);
+        File file = new File(PATH_TO_MEDIA + username + "/photo/" + multipartFile.getOriginalFilename());
         multipartFile.transferTo(file);
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         String email = authentication.getName();
-        Master master = masterRepository.findByEmail(email).orElseThrow(() -> new MasterNotFoundException(email));
-        if (!master.getPhotoLink().isEmpty()) {
-            Path path = Paths.get(master.getPhotoLink());
-            Files.delete(path);
+        if (!Objects.equals(email, "anonymousUser") && email != null) {
+            Master master = masterRepository.findByEmail(email).orElseThrow(() -> new MasterNotFoundException(email));
+            if (!master.getPhotoLink().isEmpty()) {
+                Path path = Paths.get(master.getPhotoLink());
+                Files.delete(path);
+            }
+            master.setPhotoLink(file.getAbsolutePath());
+            masterRepository.save(master);
         }
-        log.info(pathStr);
-        log.info(file.getAbsolutePath());
-        master.setPhotoLink(pathStr);
-        masterRepository.save(master);
-
         return file.getAbsolutePath();
     }
